[Design 🎨] 드롭다운 컴포넌트 디자인 반영
